requirements_clean.txt
# Core Flask and Web Framework
Flask==3.1.1
Flask-SocketIO==5.3.6
Flask-Session==0.8.0
Werkzeug==3.1.3

# Authentication and Security - Teams SSO
PyJWT==2.10.1
cryptography==45.0.6
msal==1.33.0
requests-oauthlib==2.0.0
oauthlib>=3.0.0

# HTTP and Networking
requests==2.32.4
aiohttp==3.12.15
httpx==0.28.1
httpcore==1.0.9
certifi==2025.8.3
urllib3==2.5.0

# AI and ML Libraries
groq==0.30.0
pydantic==2.11.7
pydantic_core==2.33.2

# Async and Event Handling
asyncio-happyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.9.0
eventlet==0.33.3
python-engineio==4.7.1
python-socketio==5.9.0

# Utilities and Support
python-dotenv==1.1.1
click==8.2.1
colorama==0.4.6
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0

# Data Processing
attrs==25.3.0
annotated-types==0.7.0
typing_extensions==4.14.1
typing-inspection==0.4.1
multidict==6.6.3
frozenlist==1.7.0
yarl==1.20.1

# Additional Libraries
pillow==11.3.0
propcache==0.3.2
bidict==0.23.1
greenlet==3.2.3
distro==1.9.0
dnspython==2.7.0
h11==0.16.0
six==1.17.0
sniffio==1.3.1
simple-websocket==1.1.0
wsproto==1.2.0
cachelib==0.13.0
msgspec>=0.18.6
cffi>=1.14
pycparser
idna<4,>=2.5
charset_normalizer<4,>=2
